Moved and fixed up Maybe imports
