+++ types/graphql/execution/values.d.ts
@@ -1,9 +1,9 @@
+import Maybe from "../tsutils/Maybe";
 import { GraphQLError } from "../error";
 import { GraphQLInputType, GraphQLField, GraphQLArgument } from "../type/definition";
 import { GraphQLDirective } from "../type/directives";
 import { GraphQLSchema } from "../type/schema";
 import { FieldNode, DirectiveNode, VariableDefinitionNode } from "../language/ast";
-import Maybe from "../Maybe";
 
 interface CoercedVariableValues {
     errors: ReadonlyArray<GraphQLError> | undefined;
