+++ types/graphql/utilities/valueFromAST.d.ts
@@ -1,6 +1,6 @@
+import Maybe from "../tsutils/Maybe";
 import { GraphQLInputType } from "../type/definition";
 import { ValueNode, VariableNode, ListValueNode, ObjectValueNode } from "../language/ast";
-import Maybe from "../Maybe";
 
 /**
  * Produces a JavaScript value given a GraphQL Value AST.
