+++ types/graphql/validation/rules/NoUndefinedVariables.d.ts
@@ -1,6 +1,6 @@
+import Maybe from "../../tsutils/Maybe";
 import ValidationContext from "../ValidationContext";
 import { ASTVisitor } from "../../language/visitor";
-import Maybe from "../../Maybe";
 
 export function undefinedVarMessage(varName: string, opName: Maybe<string>): string;
